CTCP-4557: Create packages subfolder and move in associated files
