+++ app/views/houseConsignment/index/items/packages/NumberOfPackagesView.scala.html
@@ -14,8 +14,8 @@
  * limitations under the License.
  *@
 
-@import controllers.houseConsignment.index.items.routes
-@import viewModels.houseConsignment.index.items.NumberOfPackagesViewModel
+@import controllers.houseConsignment.index.items.packages.routes
+@import viewModels.houseConsignment.index.items.packages.NumberOfPackagesViewModel
 @import models.MovementReferenceNumber
 @import viewModels.components.InputTextViewModel._
 @import views.html.components.{Button, ErrorSummary, InputText}
