+++ app/views/houseConsignment/index/items/packages/PackageShippingMarkView.scala.html
@@ -14,12 +14,12 @@
  * limitations under the License.
  *@
 
-@import controllers.houseConsignment.index.items.routes
+@import controllers.houseConsignment.index.items.packages.routes
 @import models.MovementReferenceNumber
 @import viewModels.components.InputCharacterCountViewModel._
 @import views.html.components.{Button, ErrorSummary, InputCharacterCount}
 @import forms.Constants.maxItemDescriptionLength
-@import viewModels.houseConsignment.index.items.PackageShippingMarksViewModel
+@import viewModels.houseConsignment.index.items.packages.PackageShippingMarksViewModel
 
 @this(
     mainTemplate: MainTemplate,
