+++ app/views/houseConsignment/index/items/packages/PackageTypeView.scala.html
@@ -14,7 +14,7 @@
  * limitations under the License.
  *@
 
-@import controllers.houseConsignment.index.items.routes
+@import controllers.houseConsignment.index.items.packages.routes
 @import models.ArrivalId
 @import models.reference.PackageType
 @import models.reference.Selectable
@@ -22,7 +22,7 @@
 @import uk.gov.hmrc.govukfrontend.views.html.components.implicits._
 @import viewModels.components.InputSelectViewModel._
 @import views.html.components.{Button, ErrorSummary, InputSelect}
-@import viewModels.houseConsignment.index.items.PackageTypeViewModel
+@import viewModels.houseConsignment.index.items.packages.PackageTypeViewModel
 
 @this(
     mainTemplate: MainTemplate,
