+++ conf/app.houseConsignment.index.items.packages.routes
@@ -1,0 +1,15 @@
+
+GET        /:arrivalId/house-consignment/:houseConsignmentIndex/items/:itemIndex/packages/:packageIndex/type-quantity          controllers.houseConsignment.index.items.packages.NumberOfPackagesController.onPageLoad(arrivalId: ArrivalId, houseConsignmentIndex: Index, itemIndex: Index, packageIndex: Index, mode: Mode = NormalMode)
+POST       /:arrivalId/house-consignment/:houseConsignmentIndex/items/:itemIndex/packages/:packageIndex/type-quantity          controllers.houseConsignment.index.items.packages.NumberOfPackagesController.onSubmit(arrivalId: ArrivalId, houseConsignmentIndex: Index, itemIndex: Index, packageIndex: Index, mode: Mode = NormalMode)
+GET        /:arrivalId/house-consignment/:houseConsignmentIndex/items/:itemIndex/packages/:packageIndex/change-type-quantity   controllers.houseConsignment.index.items.packages.NumberOfPackagesController.onPageLoad(arrivalId: ArrivalId, houseConsignmentIndex: Index, itemIndex: Index, packageIndex: Index, mode: Mode = CheckMode)
+POST       /:arrivalId/house-consignment/:houseConsignmentIndex/items/:itemIndex/packages/:packageIndex/change-type-quantity   controllers.houseConsignment.index.items.packages.NumberOfPackagesController.onSubmit(arrivalId: ArrivalId, houseConsignmentIndex: Index, itemIndex: Index, packageIndex: Index, mode: Mode = CheckMode)
+
+GET        /:arrivalId/house-consignment/:houseConsignmentIndex/items/:itemIndex/packages/:packageIndex/type              controllers.houseConsignment.index.items.packages.PackageTypeController.onPageLoad(arrivalId: ArrivalId, houseConsignmentIndex: Index, itemIndex: Index, packageIndex:Index, mode: Mode = NormalMode)
+POST       /:arrivalId/house-consignment/:houseConsignmentIndex/items/:itemIndex/packages/:packageIndex/type              controllers.houseConsignment.index.items.packages.PackageTypeController.onSubmit(arrivalId: ArrivalId, houseConsignmentIndex: Index, itemIndex: Index, packageIndex:Index, mode: Mode = NormalMode)
+GET        /:arrivalId/house-consignment/:houseConsignmentIndex/items/:itemIndex/packages/:packageIndex/change-type       controllers.houseConsignment.index.items.packages.PackageTypeController.onPageLoad(arrivalId: ArrivalId, houseConsignmentIndex: Index, itemIndex: Index, packageIndex:Index, mode: Mode = CheckMode)
+POST       /:arrivalId/house-consignment/:houseConsignmentIndex/items/:itemIndex/packages/:packageIndex/change-type       controllers.houseConsignment.index.items.packages.PackageTypeController.onSubmit(arrivalId: ArrivalId, houseConsignmentIndex: Index, itemIndex: Index, packageIndex:Index, mode: Mode = CheckMode)
+
+GET        /:arrivalId/house-consignment/:houseConsignmentIndex/items/:itemIndex/packages/:packageIndex/shipping-mark           controllers.houseConsignment.index.items.packages.PackageShippingMarkController.onPageLoad(arrivalId: ArrivalId, houseConsignmentIndex: Index, itemIndex: Index, packageIndex:Index, mode: Mode = NormalMode)
+POST       /:arrivalId/house-consignment/:houseConsignmentIndex/items/:itemIndex/packages/:packageIndex/shipping-mark           controllers.houseConsignment.index.items.packages.PackageShippingMarkController.onSubmit(arrivalId: ArrivalId, houseConsignmentIndex: Index, itemIndex: Index, packageIndex:Index, mode: Mode = NormalMode)
+GET        /:arrivalId/house-consignment/:houseConsignmentIndex/items/:itemIndex/packages/:packageIndex/change-shipping-mark    controllers.houseConsignment.index.items.packages.PackageShippingMarkController.onPageLoad(arrivalId: ArrivalId, houseConsignmentIndex: Index, itemIndex: Index, packageIndex:Index, mode: Mode = CheckMode)
+POST       /:arrivalId/house-consignment/:houseConsignmentIndex/items/:itemIndex/packages/:packageIndex/change-shipping-mark    controllers.houseConsignment.index.items.packages.PackageShippingMarkController.onSubmit(arrivalId: ArrivalId, houseConsignmentIndex: Index, itemIndex: Index, packageIndex:Index, mode: Mode = CheckMode)
